Refactored main package
