+++ src/main/java/net/vortexdata/tsqpf_plugin_antiafk/AntiAFK.java
@@ -1,8 +1,7 @@
-package net.vortexdata.tsqpf_plugin_antiafk.AntiAFK;
+package net.vortexdata.tsqpf_plugin_antiafk;
 
-import com.github.theholywaffle.teamspeak3.api.event.*;
 import net.vortexdata.tsqpf.plugins.*;
-import net.vortexdata.tsqpf_plugin_antiafk.collectors.*;
+import net.vortexdata.tsqpf_plugin_antiafk.collectors.AfkCollector;
 
 public class AntiAFK extends TeamspeakPlugin {
